+++ addOns/authhelper/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Automation Framework `diagnostics` job to start and stop plan-level recording of authentication related diagnostics.
+
 ### Changed
 - Update dependency.
 - Include the Zest statement index in the authentication diagnostics' steps.